#202: Rename remote_services → rmsvc to shorten debug log paths
