+++ .github/workflows/unit-tests.yml
@@ -19,7 +19,7 @@ jobs:
           extra-packages: libcmocka-dev python3-pytest python3-pytest-mock
 
       - name: Build object files
-        run: make src/xpath.o src/conf.o src/mem.o src/log.o src/process.o src/remote_services.o src/evdev.o
+        run: make src/xpath.o src/conf.o src/mem.o src/log.o src/process.o src/rmsvc.o src/evdev.o
 
       - name: Run C unit tests
         run: make test-c
